routes-required-services-app/requirements.txt
fastapi==0.112.3
pydantic==2.8.2
uvicorn==0.30.6
requests==2.32.3
geojson==3.1.0
shapely==2.0.6
paho-mqtt==1.6.1
